pyproject.toml
[tool.poetry]
name = "django-explorer"
version = "0.2.0"
description = "Serve local direcotry listing from django"
authors = ["dhvcc <1337kwiz@gmail.com>"]
license = "GPLv3"
include = ["LICENSE", "README.md"]
readme = "README.md"
homepage = "https://github.com/dhvcc/django_explorer"
repository = "https://github.com/dhvcc/django_explorer"
keywords = ["django", "python3", "directory", "serving", "diretory-listing"]
classifiers = [
  "Environment :: Plugins",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU Affero General Public License v3",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: JavaScript",
  "Framework :: Django",
  "Framework :: Django :: 3.0",
  "Framework :: Django :: 3.1",
  "Framework :: Django :: 3.2",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/dhvcc/django_explorer/issues"

[tool.poetry.dependencies]
python = ">=3.6.2,<4"
django = "^3"
python-magic = "^0.4.22"
pydantic = "^1.8.2"
ipython = { version = "^7.0.0", python = ">=3.7" }

[tool.poetry.dev-dependencies]
black = "^21.5b1"
isort = "^5.8.0"
flake8 = "^3.9.2"
pytest = "^6.2.4"
django-extensions = "^3.1.3"
pytest-django = "^4.3.0"
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py37', 'py38']
exclude = '''
(
    \.eggs
  | \.git
  | build
  | dist
  | venv
  | .venv
  | migrations
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip_gitignore = true
skip_glob = ['**/migrations/**', '**/.venv/**']
